Inserts test user into db
This is obviously very hackish right now, but I've
proven that I can get the test User to insert
into the database.

A painfully obvious avenue for improvement is to
not continue to expand the if/else code in SetterUtil, but, instead,
make some sort of mapper to pick the correct setter/converter,
especially, so that new setter/converters can be added dynamically
later.